RagAPI/requirements.txt
weaviate-client==4.7.1
fastapi==0.111.1